Move KryoSerialization into example.application package
