+++ sample-app/src/main/resources/akka.conf
@@ -21,7 +21,7 @@ akka {
 
     # KryoSerializableトレイトを実装するクラスは kryo でシリアライズされる
     serialization-bindings {
-      "example.adapter.KryoSerializable" = kryo
+      "example.application.KryoSerializable" = kryo
     }
   }
 }
