+++ sample-app/src/main/scala/example/application/KryoSerializable.scala
@@ -1,4 +1,4 @@
-package example.adapter
+package example.application
 
 /** kryo でシリアライズするためのマーカー
   */
